Added groups

+++ Engine/Shaders/ComputePipeline/compute_lights.glsl
@@ -30,7 +30,7 @@ layout(std430, binding = 5) restrict buffer clusterSSBO
 
 layout(std430, binding = 3) readonly buffer Omni
 {
-    vec4 lenOmni;
+    ivec4 lenOmni;
     OmniData omniData[];
 };
